refactor(studio): share the flat-inspector test render harness across panel suites
The eight flat property-panel suites each carried an identical copy of the
React act-environment setup and a local renderInto() helper. Extract both into
testRenderUtils so the harness has one definition to fix.

No behavior change: setupReactActEnvironment() registers the same
document.body cleanup each file registered inline, and renderInto() is the
same createRoot + act wrapper.
